+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Preload static indicator chart configuration data on app load
 - Use stored coordinates and switch geocoders in `ingest_missy_dataset`
 - Change cyclone and tropical storms to coastal_only
+- Do not log users out if Temperate's Climate API token is invalid
 
 ## [1.0.0] - 2018-04-03
 - MVP Release